pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alphaedge"
version = "1.0.0"
description = "Advanced Stock Market Prediction Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Aditya", email = "achyuth@alphaedge.ai"},
]
keywords = ["stock", "prediction", "machine-learning", "finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.10",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "yfinance>=0.2.36",
    "xgboost>=2.0.0",
    "torch>=2.2.0",
    "scikit-learn>=1.4.0",
    "ta>=0.11.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "joblib>=1.3.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "firebase-admin>=6.4.0",
    "google-cloud-firestore>=2.14.0",
    "PyJWT[crypto]>=2.8.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.0",
    "shap>=0.44.0",
    "gunicorn>=22.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/adityas/AlphaEdge"
Documentation = "https://docs.alphaedge.ai"
Repository = "https://github.com/adityas/AlphaEdge"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
